refactor(core): extract looksLikeSubcommand into leaf module
